backend/chroma/requirements.txt
ollama==0.4.7
chromadb>=0.6.3
langchain-ollama>=0.2.3
langchain-community>=0.0.10
PyYAML>=6.0.0
tqdm>=4.66.2